lib: switch cgroup bit-fields from signed to unsigned int
There is a problem in cgroup lib that the declearation int
can lead to -1 during the | operation.

Becasue if the field contains uninitialized garbage data,
a bit-field declared as int could interpret 0b1 as -1 due
to signed arithmetic.

By changing the type to unsigned int, the issue is avoided
since unsigned fields cannot represent negative values.

Signed-off-by: Li Wang <liwang@redhat.com>
Cc: Jin Guojie <guojie.jin@gmail.com>