testcases/kernel/mem: Move check_hugepage() + PATH_THP
These were only used in thp testcases, also after this the mem library
is no longer needed in thp tests so it's removed from the Makefile.

Also note that PATH_HUGEPAGES is defined in tst_hugepage.h which is
included by tst_test.h so we can just remove this macro.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>