memcg_control_test: Disable swapping in test cgroup
memcg_control_test_1 sets both a memory limit and a memsw/swap limit to
a single page, and then asks mem_process to allocate 2 pages in an
attempt to cause a memcg OOM kill. However, if swapping is enabled the
first page can be swapped before the second is allocated causing neither
limit to be exceeded, and no OOM kill to occur which results in a test
failure. Fix this by disabling swap in the test cgroup if it is enabled,
causing both pages to always be accounted under the same counter such
that memory.max (v2) / memory.limit_in_bytes (v1) causes an OOM kill.

Link: https://lore.kernel.org/ltp/20250326223532.3411149-1-tjmercier@google.com/
Fixes: c0b815736 ("New testcase added by Naufal.")
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: T.J. Mercier <tjmercier@google.com>