doc: Add sphinx to requirements.txt
Although that slightly prolongs creating virtualenv (and CI job), but
it's probably better to handle it as the other requirements.

Use the same sphinx and sphinx-rtd-theme versions as on readthedocs.org.
That way we get doc generation tested by CI and by local build (to avoid
incompatibility changes).

Also don't install python3-sphinx via apt in CI job to keep usage the
same as in readthedocs.org.

Suggested-by: Andrea Cervesato <andrea.cervesato@suse.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>