runtest: Move ipc into syscalls
It's not worth to keep separate runtest file for 8 very quick tests.
syscalls-ipc would be obvious runtest to move, but it contains tests for
message queues, semaphores etc. Other IPC tests are in syscalls runtest,
therefore move tests there.

Remove test description (can be visible from help and the test output,
it was also below the test not above).

Signed-off-by: Petr Vorel <pvorel@suse.cz>