lib/tst_test.c: Fix tst_brk() handling
This makes the tst_brk() handling cleaner and saner as instead of
propagating the tst_brk() result in a return value an abort flag is
introduced into the shared memory.

Now:

- All the processes but the library one that reports the results exit
  with 0
- tst_brk(TBROK, ...) increments result conters, sets the abort flag.
  and exit current process
- all other tst_brk() variants will just increments the countes and
  exits the current process

This makes the tst_brk() behavior well defined so we can now even call
tst_brk() with TFAIL and TPASS as well.

And since TBROK is supposed to exit the test immediatelly (i.e.
unrecoverable error) we are now properly doing so.

The case that main test pid called TBROK was working correctly before
this patch, since send the SIGKILL signal to he process group after we
waited for the main test pid. All that was missing is a code that sends
a signal to the main test pid in the case that TBROK was triggered by
one of it's children and now we properly kill all test processes in that
case as well.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>