lsm: fix overset attr test
LSM(s) usually handle their own internal errors in a different way,
so the right way to check if they return error, is to verify that the
common return value is -1. This is the max we can do, since errno might
vary according to the LSM implementation.

At the same time, overset attr test is _not_ checking if attr is
overset, but rather checking if attr is out-of-bounds, considering OR
operator as a valid way to generate an invalid value with
LSM_ATTR_CURRENT. This is not correct, since any OR operation using
LSM_ATTR_CURRENT will generate a valid value for the LSM(s) code. So we
remove this test that doesn't make much sense at the moment and replace
it with an "invalid attr test" instead.

Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>