commands: Remove eject test
CD-ROM is not much used device nowadays. Also the test is not in
runtest/commands nor in any other runtest command.

The main reason to remove is that this mostly test user space
implementation which barely touches kernel. Test should be in the
in user space projects (e.g. util-linux or BusyBox).

Signed-off-by: Petr Vorel <pvorel@suse.cz>