sigrelse01: Select signals based on SIGRTMIN/SIGRTMAX for musl compat
This avoids selecting signal 34 when the test is run using
musl. Signal 34 is used internally by musl as SIGSYNCCALL.
Consequently, musl's signal() will return with an error status and
errno set to EINVAL when trying to setup a signal handler for signal
34, causing the sigrelse01 test to fail.

Thanks to Pedro Falcato for suggesting using SIGRTMIN and SIGRTMAX for
this check.

Signed-off-by: Florian Schmaus <florian.schmaus@codasip.com>